Always represent feerates as f64
Mixing them up seems to have caused rounding errors.

+++ src/daemon.rs
@@ -535,7 +535,7 @@ impl Daemon {
     // Get estimated feerates for the provided confirmation targets using a batch RPC request
     // Missing estimates are logged but do not cause a failure, whatever is available is returned
     #[allow(clippy::float_cmp)]
-    pub fn estimatesmartfee_batch(&self, conf_targets: &[u16]) -> Result<HashMap<u16, f32>> {
+    pub fn estimatesmartfee_batch(&self, conf_targets: &[u16]) -> Result<HashMap<u16, f64>> {
         let params_list: Vec<Value> = conf_targets.iter().map(|t| json!([t])).collect();
 
         Ok(self
@@ -561,7 +561,7 @@ impl Daemon {
                 }
 
                 // from BTC/kB to sat/b
-                Some((*target, (feerate * 100_000f64) as f32))
+                Some((*target, feerate * 100_000f64))
             })
             .collect())
     }
