+++ src/new_index/query.rs
@@ -29,7 +29,7 @@ pub struct Query {
     chain: Arc<ChainQuery>, // TODO: should be used as read-only
     mempool: Arc<RwLock<Mempool>>,
     daemon: Arc<Daemon>,
-    cached_estimates: RwLock<(HashMap<u16, f32>, Option<Instant>)>,
+    cached_estimates: RwLock<(HashMap<u16, f64>, Option<Instant>)>,
     cached_relayfee: RwLock<Option<f64>>,
 
     #[cfg(feature = "liquid")]
@@ -143,7 +143,7 @@ impl Query {
         TransactionStatus::from(self.chain.tx_confirming_block(txid))
     }
 
-    pub fn estimate_fee(&self, conf_target: u16) -> Option<f32> {
+    pub fn estimate_fee(&self, conf_target: u16) -> Option<f64> {
         if let (ref cache, Some(cache_time)) = *self.cached_estimates.read().unwrap() {
             if cache_time.elapsed() < Duration::from_secs(FEE_ESTIMATES_TTL) {
                 return cache.get(&conf_target).copied();
@@ -159,7 +159,7 @@ impl Query {
             .copied()
     }
 
-    pub fn estimate_fee_map(&self) -> HashMap<u16, f32> {
+    pub fn estimate_fee_map(&self) -> HashMap<u16, f64> {
         if let (ref cache, Some(cache_time)) = *self.cached_estimates.read().unwrap() {
             if cache_time.elapsed() < Duration::from_secs(FEE_ESTIMATES_TTL) {
                 return cache.clone();
